mmc: host: cmdq: Check if tag info extraced from CQTERRI is valid
In the case of data errors (data CRC/timeout), we are extracting the
tag associated with the on-going command (in-case command is active)
instead of the tag associated with ongoing data. Which is wrong.
Updated the logic to use tag associated with data for data errors.

In some cases, CQTERRI is information is not reliable. So check if the
tag info provided by CQTERRI is valid or not. If not valid, complete
the errord request with any valid tag, so that all requests gets
requeued.

Without these above two check, we may use wrong tag (sometimes
completed tag, eg if CMD13 is going on CMD line, we have task field
as zero, it can be treated as tag-0 ) which can result in NULL pointer
dereference issues.

Change-Id: Ib4bf37ec5b23e88af2bdaacf6ce794b3c4785773
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>